refactor: use CompositeParametersGatherer and EmptyParametersGatherer in salesforcedx-utils-vscode
