+++ packages/salesforcedx-utils-vscode/src/commands/parameterGatherers.ts
@@ -6,6 +6,30 @@
  */
 import { CancelResponse, ContinueResponse, ParametersGatherer } from '../types';
 
+export class CompositeParametersGatherer<T> implements ParametersGatherer<T> {
+  private readonly gatherers: ParametersGatherer<any>[];
+  public constructor(...gatherers: ParametersGatherer<any>[]) {
+    this.gatherers = gatherers;
+  }
+  public async gather(): Promise<CancelResponse | ContinueResponse<T>> {
+    const aggregatedData: any = {};
+    for (const gatherer of this.gatherers) {
+      const input = await gatherer.gather();
+      if (input.type === 'CONTINUE') {
+        Object.keys(input.data).map(key => (aggregatedData[key] = input.data[key]));
+      } else {
+        return {
+          type: 'CANCEL'
+        };
+      }
+    }
+    return {
+      type: 'CONTINUE',
+      data: aggregatedData
+    };
+  }
+}
+
 export class EmptyParametersGatherer implements ParametersGatherer<{}> {
   public async gather(): Promise<CancelResponse | ContinueResponse<{}>> {
     return { type: 'CONTINUE', data: {} };
