+++ packages/salesforcedx-vscode-core/src/commands/aliasList.ts
@@ -5,8 +5,9 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { Command, SfCommandBuilder } from '@salesforce/salesforcedx-utils';
+import { EmptyParametersGatherer } from '@salesforce/salesforcedx-utils-vscode';
 import { nls } from '../messages';
-import { EmptyParametersGatherer, SfCommandlet, SfCommandletExecutor, SfWorkspaceChecker } from './util';
+import { SfCommandlet, SfCommandletExecutor, SfWorkspaceChecker } from './util';
 
 export class AliasList extends SfCommandletExecutor<{}> {
   public build(data: {}): Command {
