+++ packages/salesforcedx-vscode-core/src/commands/source/viewChanges.ts
@@ -4,7 +4,8 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { EmptyParametersGatherer, SfCommandlet, SfWorkspaceChecker } from '../util';
+import { EmptyParametersGatherer } from '@salesforce/salesforcedx-utils-vscode';
+import { SfCommandlet, SfWorkspaceChecker } from '../util';
 import { SourceTrackingGetStatusExecutor } from './sourceTrackingGetStatusExecutor';
 
 const workspaceChecker = new SfWorkspaceChecker();
