+++ packages/salesforcedx-vscode-core/src/commands/templates/visualforceGenerateComponent.ts
@@ -5,11 +5,10 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { DirFileNameSelection, LocalComponent } from '@salesforce/salesforcedx-utils-vscode';
+import { CompositeParametersGatherer, DirFileNameSelection, LocalComponent } from '@salesforce/salesforcedx-utils-vscode';
 import { TemplateType, VisualforceComponentOptions } from '@salesforce/templates';
 import { nls } from '../../messages';
 import {
-  CompositeParametersGatherer,
   MetadataTypeGatherer,
   SelectFileName,
   SelectOutputDir,
