+++ packages/salesforcedx-vscode-core/src/commands/util/index.ts
@@ -19,9 +19,7 @@ export { formatException } from './formatException';
 export { LibraryPathsGatherer } from './libraryPathsGatherer';
 export { OverwriteComponentPrompt } from './overwriteComponentPrompt';
 export {
-  CompositeParametersGatherer,
   DemoModePromptGatherer,
-  EmptyParametersGatherer,
   FilePathGatherer,
   FileSelection,
   FileSelector,
