+++ packages/salesforcedx-vscode-core/test/jest/commands/templates/lightningGenerateLwc.test.ts
@@ -5,10 +5,10 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { CompositeParametersGatherer } from '@salesforce/salesforcedx-utils-vscode';
 import { lightningGenerateLwc } from '../../../../src/commands/templates';
 import { OverwriteComponentPrompt } from '../../../../src/commands/util/overwriteComponentPrompt';
 import {
-  CompositeParametersGatherer,
   MetadataTypeGatherer,
   SelectFileName,
   SelectOutputDir,
